+++ src/loadpathutil.h
@@ -1,1 +1,0 @@
-void loadPathHack (char buf[256], void *addr);